+++ include/QuadratureLfo.h
@@ -77,10 +77,11 @@ public:
 
 	void tick( float *l, float *r )
 	{
-		*l = std::sin(m_phase);
-		*r = std::sin(m_phase + m_offset);
+		*l = sinf( m_phase );
+		*r = sinf( m_phase + m_offset );
 		m_phase += m_increment;
-		while (m_phase >= numbers::tau)	{ m_phase -= numbers::tau; }
+
+		while (m_phase >= numbers::tau)	{	m_phase -= numbers::tau; }
 	}
 
 private:
